requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
bert-serving-client==1.10.0
bert-serving-server==1.10.0
boto3==1.12.42
botocore==1.15.42
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
datasketch==1.5.1
deepmatcher===0.1.1-pt
distlib==0.3.0
fasttext==0.9.1
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
jmespath==0.9.5
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.2
keras-self-attention==0.42.0
keras-word-char-embd==0.20.0
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.1.3
mock==4.0.2
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo-spark==0.1.dev0
pyparsing==2.4.6
python-dateutil==2.8.1
pytorch-transformers==1.2.0
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.41
scipy==1.4.1
seaborn==0.10.0
sentence-transformers==0.2.6.1
sentencepiece==0.1.85
six==1.14.0
strsimpy==0.1.4
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==1.13.0rc1
tensorflow-estimator==1.15.1
termcolor==1.1.0
textdistance==4.2.0
tokenizers==0.5.2
tqdm==4.45.0
transformers==2.8.0
urllib3==1.25.8
virtualenv==20.0.5
Werkzeug==0.16.1
wrapt==1.12.1
zipp==3.0.0
